changing readOBJ to support more generic OBJ file. | a few minor optimizations

+++ src/unwrapper.h
@@ -51,7 +51,7 @@ private:
 	// output:
 	// _adj_count: number of neighbours each triangle has
 	// _triNeighbor: contains the index of each triangle neighbors
-	static void findTriangleNeighbors(const Mesh3D& _mesh, std::vector<unsigned int>& _adj_count, std::vector<int>& _triNeighbor);
+	static void findTriangleNeighbors(const Mesh3D& _mesh, std::vector<size_t>& _adj_count, std::vector<int>& _triNeighbor);
 
 	// Checks if 2D vector v1 (from v1a to v1b) intersects with v2 or not
 	static bool vectorIntersec(const Vector2f& _v1a, const Vector2f& _v1b, const Vector2f& _v2a, const Vector2f& _v2b);
